requirements.txt
charset_normalizer==3.4.2
json5==0.12.0
jsonschema==4.24.0
langchain_community==0.3.27
langchain_core==0.3.68
langchain_experimental==0.3.4
loguru==0.7.3
markdownify==1.1.0
mcp==1.11.0
openai==1.95.1
Pillow==11.3.0
pydantic==2.11.7
PyGithub==2.6.1
PyGithub==2.6.1
python-dotenv==1.1.1
PyYAML==6.0.2
PyYAML==6.0.2
readabilipy==0.3.0
Requests==2.32.4
setuptools==80.9.0
tavily_python==0.7.5
